move post without pre to the end of the trace

+++ analyzer/trace/traceElementAtomic.go
@@ -113,8 +113,8 @@ func (at *traceElementAtomic) getTpost() int {
  * Returns:
  *   int: The timer of the element
  */
-func (at *traceElementAtomic) getTsort() float32 {
-	return float32(at.tpost)
+func (at *traceElementAtomic) getTsort() int {
+	return at.tpost
 }
 
 /*
